requirements.txt
Flask~=3.0.3
python-dotenv~=1.0.1
SQLAlchemy~=2.0.35
pandas~=2.2.3
pycryptodome~=3.20.0